Main file rename and archive of terminal version
