+++ .gitignore
@@ -1,3 +1,5 @@
 /inputs/
 /outputs/
 /Temp_files/
+/terminal_archive/inputs
+/terminal_archive/outputs
